Fixed Visual Bug with button
The issue that was causing this bug was multiple pygame.display.update()'s being called at once in main. Only one should ever be called in a single frame. I made a help function that allows the animation of the algorithm to be displayed. Now whenever you call A*, use animated_draw, not draw.